pyproject.toml
[tool.poetry]
name = "ubend_pl"
version = "1.0"
description = ""
authors = ["Danil Muranov <danil5678444@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
pandas = "^2.0.2"
scikit-learn = "^1.2.0"
matplotlib = "^3.7.1"
seaborn = "^0.12.2"
hydra-core = "^1.3.2"
pytest = "^7.4.0"
torch = {version = "*", source = "torch_cpu"}
psutil = "^5.9.8"

[[tool.poetry.source]]
name = "torch_cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "supplemental"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[tool.poetry.scripts]
make_dataset = "ubend_pl.data.make_dataset:make_dataset"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 80
include = '\.pyi?$'
exclude = '''
/(
	\.git
| \.hg
| \.mypy_cache
| \.tox
| \.venv
| _build
| buck-out
| build   
)/ 
'''
